minimega: rename gui to web.
And a few other cleanups. Ready for beta testing...
